# [Not for Land][FSDP] Only sync free event if needs unshard