@awgu 2023-04-05
@rohan-varma I do not have a conclusive understanding of the memory effect here, but this PR regresses `limit_all_gathers=True`. **We should not land it.**

I suspect the issue is that with this PR, the reduce-scatter allocations become "over-prefetched" in the same way the all-gather allocations did. Before this PR, `limit_all_gathers=True` blocks for the unshard even if not all-gathering, which in turn prevents reduce-scatter over-allocation.

Empirically, I tested the two regimes: fast CPU thread that is (1) GPU compute bound and (2) GPU communication bound.

**GPU Compute Bound**
- Baseline `limit_all_gathers=False`:
    - Time per batch: 247.559 ms
    - Peak active bytes: 38555 MB
    - Peak reserved bytes: 38806 MB
- Baseline `limit_all_gathers=True`:
    - Time per batch: 236.484 ms
    - Peak active bytes: 7164 MB
    - Peak reserved bytes: 8032 MB
- New `limit_all_gathers=True`:
    - Time per batch: 235.161 ms
    - Peak active bytes: 10292 MB
    - Peak reserved bytes: 10900 MB

**GPU Communication Bound**
- Baseline `limit_all_gathers=False`:
    - Time per batch: 827.490 ms
    - Peak active bytes: 38937 MB
    - Peak reserved bytes: 39186 MB 
- Baseline `limit_all_gathers=True`:
    - Time per batch: 815.192 ms
    - Peak active bytes: 8691 MB
    - Peak reserved bytes: 9178 MB
- New `limit_all_gathers=True`:
    - Time per batch: 809.414 ms
    - Peak active bytes: 11820 MB
    - Peak reserved bytes: 12428 MB

We see that the new `limit_all_gathers=True` is marginally faster for the communication bound regime (due to scheduling backward ops slightly sooner), but it shows a nontrivial memory regression: 7164 MB -> 10292 MB and 8691 MB -> 11820 MB peak active memory.